add 5 lesson code

+++ server.py
@@ -1,30 +1,33 @@
+import math
+import random
 import socket
-
-import pygame
-from sqlalchemy import Column, Integer, String
+import time
+import psycopg2
 from sqlalchemy import create_engine
 from sqlalchemy.orm import declarative_base
+from sqlalchemy import Column, Integer, String
 from sqlalchemy.orm import sessionmaker
+import pygame
+from russian_names import RussianNames
 
-main_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)  # настраиваем сокет
-main_socket.setsockopt(socket.IPPROTO_TCP, socket.TCP_NODELAY, 1)  # отключаем пакетирование
-main_socket.bind(("localhost", 10000))  # ip и порт привязываем к порту
-main_socket.setblocking(False)  # непрерывность, не ждем ответа
-main_socket.listen(5)  # прослушка входящих соединений, 5 одновременных подключений
-print("Сокет создался!")
-
-engine = create_engine(
-    "postgresql+psycopg2://postgres:ASDFgh123!@localhost/postgres"
-)
+engine = create_engine("postgresql+psycopg2://postgres:ASDFgh123!@localhost/postgres")
 Session = sessionmaker(bind=engine)
 Base = declarative_base()
 s = Session()
-pygame.init()
 
+# Настройка констант для pygame
+pygame.init()
 WIDHT_ROOM, HEIGHT_ROOM = 4000, 4000
 WIDHT_SERVER, HEIGHT_SERVER = 300, 300
-FPS = 60
+FPS = 100
+colors = ['Maroon', 'DarkRed', 'FireBrick', 'Red', 'Salmon', 'Tomato', 'Coral', 'OrangeRed', 'Chocolate', 'SandyBrown',
+          'DarkOrange', 'Orange', 'DarkGoldenrod', 'Goldenrod', 'Gold', 'Olive', 'Yellow', 'YellowGreen', 'GreenYellow',
+          'Chartreuse', 'LawnGreen', 'Green', 'Lime', 'SpringGreen', 'MediumSpringGreen', 'Turquoise',
+          'LightSeaGreen', 'MediumTurquoise', 'Teal', 'DarkCyan', 'Aqua', 'Cyan', 'DeepSkyBlue',
+          'DodgerBlue', 'RoyalBlue', 'Navy', 'DarkBlue', 'MediumBlue']
+MOBS_QUANTITY = 25
 
+# Создание окна сервера
 screen = pygame.display.set_mode((WIDHT_SERVER, HEIGHT_SERVER))
 pygame.display.set_caption("Сервер")
 clock = pygame.time.Clock()
@@ -35,11 +38,11 @@ def find(vector: str):
     for num, sign in enumerate(vector):
         if sign == "<":
             first = num
-        if sign == "<" and first is not None:
+        if sign == ">" and first is not None:
             second = num
-            result = list(map(float, vector[first + 1: second].split(",")))
+            result = list(map(float, vector[first + 1:second].split(",")))
             return result
-    return " "
+    return ""
 
 
 def find_color(info: str):
@@ -54,6 +57,7 @@ def find_color(info: str):
     return ""
 
 
+# Декларативный класс таблицы игроков
 class Player(Base):
     __tablename__ = "gamers"
     id = Column(Integer, primary_key=True, autoincrement=True)
@@ -66,15 +70,16 @@ class Player(Base):
     abs_speed = Column(Integer, default=1)
     speed_x = Column(Integer, default=0)
     speed_y = Column(Integer, default=0)
-    color = Column(String(250), default="red")
+    color = Column(String(250), default="red")  # Добавили цвет
     w_vision = Column(Integer, default=800)
-    h_vision = Column(Integer, default=600)
+    h_vision = Column(Integer, default=600)  # Добавили размер квадрат
 
     def __init__(self, name, address):
         self.name = name
         self.address = address
 
 
+# Локальный класс таблицы игроков
 class LocalPlayer:
     def __init__(self, id, name, sock, addr):
         self.id = id
@@ -94,8 +99,25 @@ class LocalPlayer:
         self.h_vision = 600
 
     def update(self):
-        self.x += self.speed_x
-        self.y += self.speed_y
+        # х координата
+        if self.x - self.size <= 0:  # Если игрок вылазит за левую стенку
+            if self.speed_x >= 0:  # Но при этом двигается право
+                self.x += self.speed_x  # то двигаем его
+        elif self.x + self.size >= WIDHT_ROOM:  # Если игрок вылазит за правую стенку
+            if self.speed_x <= 0:  # Но при этом двигается влево
+                self.x += self.speed_x  # то двигаем его
+        else:  # Если игрок находится в границе комнаты
+            self.x += self.speed_x
+
+        # y координата
+        if self.y - self.size <= 0:  # Если игрок вылазит за левую стенку
+            if self.speed_y >= 0:  # Но при этом двигается право
+                self.y += self.speed_y  # то двигаем его
+        elif self.y + self.size >= HEIGHT_ROOM:  # Если игрок вылазит за правую стенку
+            if self.speed_y <= 0:  # Но при этом двигается влево
+                self.y += self.speed_y  # то двигаем его
+        else:  # Если игрок находится в границе комнаты
+            self.y += self.speed_y
 
     def change_speed(self, vector):
         vector = find(vector)
@@ -136,45 +158,78 @@ class LocalPlayer:
 
 Base.metadata.create_all(engine)
 players = {}
+
+main_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)  # Настраиваем сокет
+main_socket.setsockopt(socket.IPPROTO_TCP, socket.TCP_NODELAY, 1)  # Отключаем пакетирование
+main_socket.bind(("localhost", 10000))  # IP и порт привязываем к порту
+main_socket.setblocking(False)  # Непрерывность, не ждём ответа
+main_socket.listen(5)  # Прослушка входящих соединений, 5 одновременных подключений
+print("Сокет создался")
+
+# Создание мобов
+names = RussianNames(count=MOBS_QUANTITY * 2, patronymic=False, surname=False, rare=True)
+names = list(set(names))  # Список неповторяющихся имён
+
+for x in range(MOBS_QUANTITY):
+    server_mob = Player(names[x], None)
+    server_mob.color = random.choice(colors)
+    server_mob.x, server_mob.y = random.randint(0, WIDHT_ROOM), random.randint(0, HEIGHT_ROOM)
+    server_mob.speed_x, server_mob.speed_y = random.randint(-1, 1), random.randint(-1, 1)
+    server_mob.size = random.randint(10, 100)
+    s.add(server_mob)
+    s.commit()
+    local_mob = LocalPlayer(server_mob.id, server_mob.name, None, None).load()
+    players[server_mob.id] = local_mob  # Записываем всех мобов в словарь
+
+tick = -1
 server_works = True
 while server_works:
     clock.tick(FPS)
-    try:
-        # проверяем желающих войти в игру
-        new_socket, addr = main_socket.accept()  # принимаем входящие
-        print('Подключился', addr)
-        new_socket.setblocking(False)
-        login = new_socket.recv(1024).decode()
-        player = Player("Имя", addr)
+    tick += 1
+    if tick % 200 == 0:
+        try:
+            # проверяем желающих войти в игру
+            new_socket, addr = main_socket.accept()  # принимаем входящие
+            print('Подключился', addr)
+            new_socket.setblocking(False)
+            login = new_socket.recv(1024).decode()
+            player = Player("Имя", addr)
 
-        if login.startswith("color"):
-            data = find_color(login[6:])
-            player.name, player.color = data
+            if login.startswith("color"):
+                data = find_color(login[6:])
+                player.name, player.color = data
 
-        s.merge(player)
-        s.commit()
+            s.merge(player)
+            s.commit()
 
-        addr = f'({addr[0]},{addr[1]})'
-        data = s.query(Player).filter(Player.address == addr)
-        for user in data:
-            player = LocalPlayer(user.id, "Имя", new_socket, addr).load()
-            players[user.id] = player
+            addr = f'({addr[0]},{addr[1]})'
+            data = s.query(Player).filter(Player.address == addr)
+            for user in data:
+                player = LocalPlayer(user.id, "Имя", new_socket, addr).load()
+                players[user.id] = player
 
-    except BlockingIOError:
-        pass
+        except BlockingIOError:
+            pass
 
     # Считываем команды игроков
     for id in list(players):
-        try:
-            data = players[id].sock.recv(1024).decode()
-            print("Получил", data)
-            players[id].change_speed(data)
-        except:
-            pass
+        if players[id].sock is not None:
+            try:
+                data = players[id].sock.recv(1024).decode()
+                players[id].change_speed(data)
+            except:
+                pass
+        else:
+            if tick % 400 == 0:
+                vector = f"<{random.randint(-1, 1)},{random.randint(-1, 1)}>"
+                players[id].change_speed(vector)  # Случайный вектор для мобов
 
     # Определим, что видит каждый игрок
+    # + съедение друг друга
+    # bacteries = {}
     visible_bacteries = {}
     for id in list(players):
+        # bacteries[id] = s.get(Player, id)  # Заполняем словарь
         visible_bacteries[id] = []
 
     pairs = list(players.items())
@@ -188,25 +243,39 @@ while server_works:
 
             # i-й игрок видит j-того
             if abs(dist_x) <= hero_1.w_vision // 2 + hero_2.size and abs(dist_y) <= hero_1.h_vision // 2 + hero_2.size:
-                # Подготовим данные к добавлению в список
-                x_ = str(round(dist_x))
-                y_ = str(round(dist_y))  # временные
-                size_ = str(round(hero_2.size))
-                color_ = hero_2.color
+                # Проверка может ли 1-й съесть 2-го игрока
+                distance = math.sqrt(dist_x ** 2 + dist_y ** 2)
+                if distance <= hero_1.size and hero_1.size > 1.1 * hero_2.size:
+                    # Меняем радиус первого игрока
+                    hero_2.size, hero_2.speed_x, hero_2.speed_y = 0, 0, 0
 
-                data = x_ + " " + y_ + " " + size_ + " " + color_
-                visible_bacteries[hero_1.id].append(data)
+                if hero_1.address is not None:
+                    # Подготовим данные к добавлению в список
+                    x_ = str(round(dist_x))
+                    y_ = str(round(dist_y))  # временные
+                    size_ = str(round(hero_2.size))
+                    color_ = hero_2.color
+
+                    data = x_ + " " + y_ + " " + size_ + " " + color_
+                    visible_bacteries[hero_1.id].append(data)
 
             # j-й игрок видит i-того
             if abs(dist_x) <= hero_2.w_vision // 2 + hero_1.size and abs(dist_y) <= hero_2.h_vision // 2 + hero_1.size:
-                # Подготовим данные к добавлению в список
-                x_ = str(round(-dist_x))
-                y_ = str(round(-dist_y))  # временные
-                size_ = str(round(hero_1.size))
-                color_ = hero_1.color
+                # Проверка может ли 2-й съесть 1-го игрока
+                distance = math.sqrt(dist_x ** 2 + dist_y ** 2)
+                if distance <= hero_2.size and hero_2.size > 1.1 * hero_1.size:
+                    # Меняем радиус второго игрока
+                    hero_1.size, hero_1.speed_x, hero_1.speed_y = 0, 0, 0
 
-                data = x_ + " " + y_ + " " + size_ + " " + color_
-                visible_bacteries[hero_2.id].append(data)
+                if hero_2.address is not None:
+                    # Подготовим данные к добавлению в список
+                    x_ = str(round(-dist_x))
+                    y_ = str(round(-dist_y))  # временные
+                    size_ = str(round(hero_1.size))
+                    color_ = hero_1.color
+
+                    data = x_ + " " + y_ + " " + size_ + " " + color_
+                    visible_bacteries[hero_2.id].append(data)
 
     # Формируем ответ каждой бактерии
     for id in list(players):
@@ -214,15 +283,25 @@ while server_works:
 
     # Отправляем статус игрового поля
     for id in list(players):
-        try:
-            players[id].sock.send(visible_bacteries[id].encode())
-        except:
-            players[id].sock.close()
+        if players[id].sock is not None:
+            try:
+                players[id].sock.send(visible_bacteries[id].encode())
+            except:
+                players[id].sock.close()
+                del players[id]
+                # Так же удаляем строчку из БД
+                s.query(Player).filter(Player.id == id).delete()
+                s.commit()
+                print("Сокет закрыт")
+
+    # Чистим список от отвалившихся игроков
+    for id in list(players):
+        if players[id].errors >= 500 or players[id].size == 0:
+            if players[id].sock is not None:
+                players[id].sock.close()
             del players[id]
-            # Так же удаляем строчку из БД
             s.query(Player).filter(Player.id == id).delete()
             s.commit()
-            print("Сокет закрыт")
 
     # Отрисовываем серверное окно
     for event in pygame.event.get():
@@ -240,6 +319,9 @@ while server_works:
     for id in list(players):
         player = players[id]
         players[id].update()
+        if tick % 300 == 0:
+            players[id].sync()
+
     pygame.display.update()
 
 pygame.quit()
